+++ synapse-core/src/main/java/de/otto/synapse/annotation/MessageQueueConsumer.java
@@ -1,4 +1,4 @@
-package de.otto.synapse.annotation.messagequeue;
+package de.otto.synapse.annotation;
 
 import de.otto.synapse.endpoint.receiver.MessageQueueReceiverEndpoint;
 import de.otto.synapse.message.Message;