+++ synapse-examples/example-messagequeue-consumer/src/main/java/de/otto/synapse/example/consumer/ExampleConsumer.java
@@ -1,6 +1,6 @@
 package de.otto.synapse.example.consumer;
 
-import de.otto.synapse.annotation.messagequeue.MessageQueueConsumer;
+import de.otto.synapse.annotation.MessageQueueConsumer;
 import de.otto.synapse.example.consumer.payload.BananaPayload;
 import de.otto.synapse.example.consumer.payload.ProductPayload;
 import de.otto.synapse.example.consumer.state.BananaProduct;
@@ -9,7 +9,6 @@ import de.otto.synapse.state.StateRepository;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import software.amazon.awssdk.services.sqs.SqsAsyncClient;
 
 import static de.otto.synapse.example.consumer.state.BananaProduct.bananaProductBuilder;
 import static java.lang.String.valueOf;